feat!: split implementation in two classes
The "_sync" folders were generated automatically
executing the command "unasync qvapay tests".

The code that is added in the "_async" folders are
automatically transformed.

So every time to make a change you must run the
command "unasync qvapay tests" to regenerate
the folders "_sync" with the synchronous version
of the implementation.